requirements.txt
grpcio==1.67.1
grpcio-tools==1.67.1
grpcio-health-checking==1.67.1
grpcio-reflection==1.67.1
protobuf==5.29.3
python-consul==1.1.0
python-dotenv==1.0.1
python-ulid==3.0.0
minio==7.2.15
pymilvus==2.5.4
insightface==0.7.3
opencv-python-headless==4.11.0.86
numpy==1.26.4
tqdm==4.67.1
requests==2.32.3
prettytable==3.15.1
# onnxruntime==1.17.3
onnxruntime==1.22.0
# dotenv==
# nats
# opencv

